refactor(transfer): 필드 및 메서드 타입 Instant에서 LocalDateTime으로 변경

+++ transfer/model/src/main/java/org/yechan/remittance/transfer/Ledger.java
@@ -1,12 +1,13 @@
 package org.yechan.remittance.transfer;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record Ledger(
     Long ledgerId,
     Long transferId,
     Long accountId,
-    Long amount,
+    BigDecimal amount,
     LedgerSideValue side,
     LocalDateTime createdAt
 ) implements LedgerModel {
